refactor: move controller's routes in a dedicated file
